ajax working

+++ db/migrate/20150910153054_create_characters.rb
@@ -2,8 +2,8 @@ class CreateCharacters < ActiveRecord::Migration
   def change
     create_table :characters do |t|
       t.string :name
-      t.decimal :x_coordinate, precision: 4
-      t.decimal :y_coordinate, precision: 4
+      t.float :x_coordinate, precision: 4
+      t.float :y_coordinate, precision: 4
       t.references :puzzles
       t.timestamps null: false
     end
